FE-464 Split tenant configs, context, and default configs into separate files
